feat: extract common homepage logic to reuse
