+++ src/domain/home/organisms/desktop-home/desktop-home.styles.ts
@@ -1,9 +1,39 @@
 import styled from '@emotion/styled';
 
-const StyledWrapper = styled.div``;
+import { respondToMax } from '@infrastructure';
+
+const StyledWrapper = styled.div`
+  width: 100%;
+  height: 100%;
+  display: flex;
+  align-items: center;
+  flex-direction: column;
+  justify-content: center;
+`;
+
+const StyledSubHeaderWrapper = styled.div`
+  display: flex;
+  align-items: center;
+  flex-direction: column;
+  justify-content: center;
+  margin-top: ${({ theme }) => theme.spacing(3)};
+
+  > .MuiButton-root {
+    margin-top: ${({ theme }) => theme.spacing(12)};
+  }
+
+  ${({ theme }) => respondToMax.xmobile`
+    margin-top: ${theme.spacing(4)};
+
+    > .MuiButton-root {
+      margin-top: ${theme.spacing(10)};
+    }
+  `}
+`;
 
 const S = {
   StyledWrapper,
+  StyledSubHeaderWrapper,
 };
 
 export default S;
