+++ src/domain/home/organisms/desktop-home/desktop-home.types.ts
@@ -1,1 +1,4 @@
-export interface IDesktopHomeProps {}
+export interface IDesktopHomeProps {
+  clientTypeId: number | undefined;
+  isSignedIn: boolean;
+}
