+++ src/domain/home/organisms/mobile-home/mobile-home.types.ts
@@ -1,1 +1,3 @@
-export interface IMobileHomeProps {}
+export interface IMobileHomeProps {
+  isSignedIn: boolean;
+}
